+++ CHANGELOG.md
@@ -23,6 +23,10 @@ get ports from config file.
 
 - Add limit checking per Josh's recommended ranges.
 
+- MTPviewer: If RCFs are missing (RC dir is empty), or other errors that cause
+prfiles not to be generated, code will catch this and display a message on the
+profile plot rather than pop up a warning box.
+
 ## [0.2] - 2022-08-29 - ACCLIP FF03 TEST
 
 This version of the software will be flown on ACCLIP FF03 from Anchorage to RMMA